Redis receiver: Expand README, add obsreport metrics.
