+++ receiver/redisreceiver/redis_runnable_test.go
@@ -43,3 +43,4 @@ func (c *fakeMetricsConsumer) ConsumeMetricsData(_ context.Context, md consumerd
 	c.md = md
 	return nil
 }
+
